requirements.txt
requests
python-dotenv
google-generativeai
google-adk